requirements.txt
py4j==0.10.9.2
pyspark==3.2.0
PyYAML==6.0.1
